requirements.txt
APScheduler==3.10.4
croniter==2.0.1
requests==2.31.0
python-dotenv==1.0.0
passlib==1.7.4
SQLAlchemy>=2.0,<3

# FastAPI Stack (Migration)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
PyJWT[crypto]>=2.8.0
python-multipart>=0.0.6
pydantic>=2.0
pydantic-settings>=2.0
email-validator>=2.0.0
httpx>=0.25.0
aiosqlite>=0.19.0
greenlet>=3.0.0

# Dev/Test
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=5.0.0
